Array support
Add ability to define fields as arrays, when the database supports it.

An array in Go is always nullable, so there is no pointer to array if
the field is nullable on the SQL side. I've gone with the Go
convention that it's OK to initialize an empty array as nil even if the
array is defined as not nullable in the model.

When reviewing, please consider wether changes are in the right place.
The codebase is new to me and there are few comments to guide
the developer. Making this change felt like a blind man touching
an elephant.

Known limitation: lib/pq only supports a few types of array elements.
Notably it doesn't support (u)int but instead (u)int32 and -64.

This change adds the first test specifically for Postgres because SQLite
does not support arrays. The test coverage can be improved: I'd like
some feedback before I put in extra effort.

A follow-up on this would be to add ways to atomically append to arrays.

Change-Id: If773dc1c900b31dfe321d3d0172361b98c79451d