testdata: use standard go tests for testing generated code
It was very hard to develop new tests / dialects, as the tests were executed based on go code which can not be compiled, only template.

The final state was available only during the test in /tmp directory.

This change makes the development experience more smooth:

 * The tests are separated and real unit tests
 * CI always generates the dbx.go files
 * As long as you generated those files locally, you can run all the tests as local tests

Note: Jenkins build target is removed, as we need the generated source for building the tests --> and it's included in the unit tests anyway...

Change-Id: I4e3178f2bdff8611786082513ec661453c4ad5a5