sql: optimize Spanner tuple comparison
Spanner used previously to convert (A, B, C) > (?, ?, ?) into:

   (A > ?) OR
   (A = ? AND B > ?) OR
   (A = ? AND B = ? AND C > ?)

This causes significant number of arguments to be passed to the
database. This can be implemented with less arguments via:

   A > ? OR (A = ? AND (B > ? OR (B = ? AND C > ?)))

This potentially could use positional arguments to avoid duplication of
the arguments altogether, however dbx in the current state makes it
difficult.

Change-Id: I0cb7700e3be4b44c1b3db041b4edd2c95fe236c0