testdata: move test files here and generate data for all
This ends up creating a lot of code, however it's probably worthwhile to
keep track all of the generated features, so it's easier to spot code
breaking.

Reorganize code such that:

  testdata/good/*.dbx - contains examples to generate code for
  testdata/bad/*.dbx - contains examples that should fail
  testdata/generated/*/* - generated code
  testdata/*_test.go - contains test code for the above

This makes it more obvious what is generated code and which is written
manually. It also avoids the `gen.go` duplication across many places.

Change-Id: I8891e8c611c915db82a3c98f5f62b49008b20903