templates: avoid errors.Join where possible
This might be a more controversial change, since our usual approach is
to take anything that might be a relevant error and combine them all
together in an error group. But in certain contexts, I feel like we can
get away with something simpler: keep the first error indicating
trouble, and throw the rest away.

For example, in the context of querying a database, if we get an error
from rows.Err() and then another from rows.Close(), it is extremely
likely that the first error needs to be resolved before we care about
the second. Of course it's possible to come up with a situation where
the .Close() error is the only relevant one, but the situation would
have to be pretty contrived.

In this change, errors from rows.Scan() or value decoding take
precedence, and will be returned alone even if there might be errors
from rows.Err() and/or rows.Close(). If there is no error from
rows.Scan() or value decoding, errors from rows.Err() come next. If
there is no error from rows.Err(), we finally return any errors from
rows.Close().

All of this will serve to avoid duplicate errors (since it's possible
for the same error to be returned from .Err() and .Close()) and having
single errors wrapped in a group (as might be returned from
errors.Join()).

Change-Id: I68a793e8fed373dc2e2e3689ef0aa6b1e3c14d75